+++ src/main/java/org/apache/accumulo/access/AccessEvaluator.java
@@ -18,9 +18,6 @@
  */
 package org.apache.accumulo.access;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
@@ -153,32 +150,7 @@ public interface AccessEvaluator {
    *
    */
   static AccessEvaluator of(Collection<Authorizations> authorizationSets) {
-    List<AccessEvaluatorImpl> evaluators = new ArrayList<>(authorizationSets.size());
-    for (Authorizations authorizations : authorizationSets) {
-      evaluators.add(new AccessEvaluatorImpl(authorizations));
-    }
-
-    return new AccessEvaluator() {
-      @Override
-      public boolean canAccess(String accessExpression) throws IllegalAccessExpressionException {
-        return canAccess(accessExpression.getBytes(UTF_8));
-      }
-
-      @Override
-      public boolean canAccess(byte[] accessExpression) throws IllegalAccessExpressionException {
-        for (AccessEvaluatorImpl evaluator : evaluators) {
-          if (!evaluator.canAccess(accessExpression)) {
-            return false;
-          }
-        }
-        return true;
-      }
-
-      @Override
-      public boolean canAccess(AccessExpression accessExpression) {
-        return canAccess(accessExpression.getExpression());
-      }
-    };
+    return new MultiAccessEvaluatorImpl(authorizationSets);
   }
 
   /**
